fix an issue with timings that was causing music to be really sped up

+++ desktop/src/main.rs
@@ -22,7 +22,7 @@ fn main() {
     skip_bios = false;
   }
 
-  let audio_buffer: Arc<Mutex<VecDeque<f32>>> = Arc::new(Mutex::new(VecDeque::new()));
+  let audio_buffer: Arc<Mutex<VecDeque<i16>>> = Arc::new(Mutex::new(VecDeque::new()));
 
   let bios7_file = "../bios7.bin";
   let bios9_file = "../bios9.bin";
@@ -61,7 +61,6 @@ fn main() {
     frame_finished = false;
 
     // render stuff
-    // frontend.push_samples(bus.arm7.apu.audio_samples.drain(..).collect());
     frontend.render(&mut bus.gpu);
     frontend.handle_events(bus);
   }
